+++ test/unit/list_user_test.rb
@@ -59,82 +59,47 @@ class ListUserTest < ActiveSupport::TestCase
   end
 
   test 'getMonthsBetween returns [] if last day after first day' do
-    firstDay = Date.new(2012, 3, 29)
-    lastDay = Date.new(2012, 3, 28)
+    first_day = Date.new(2012, 3, 29)
+    last_day = Date.new(2012, 3, 28)
 
-    # TODO: Since ListUser.months_in_timespan got changed this assert need repair
-    # assert_equal [], ListUser.months_in_timespan(firstDay..lastDay).map(&:month)
+    assert_equal [], DateTools.months_in_time_span(first_day..last_day)
   end
 
   test 'getMonthsBetween returns [3] if both days in march 2012 and equal' do
-    firstDay = Date.new(2012, 3, 27)
-    lastDay = Date.new(2012, 3, 27)
+    first_day = Date.new(2012, 3, 27)
+    last_day = Date.new(2012, 3, 27)
 
-    # TODO: Since ListUser.months_in_timespan got changed this assert need repair
-    # assert_equal [3], ListUser.months_in_timespan(firstDay..lastDay).map(&:month)
+    expected = [3]
+    current = months_numbers_in_time_span(first_day, last_day)
+    assert_equal expected, current.flatten.uniq
   end
 
-  test 'getMonthsBetween returns [3] if both days in march 2012 and different' do
-    firstDay = Date.new(2012, 3, 27)
-    lastDay = Date.new(2012, 3, 28)
+  test 'months_in_time_span returns [3] if both days in march 2012 and different' do
+    first_day = Date.new(2012, 3, 27)
+    last_day = Date.new(2012, 3, 28)
 
-    # TODO: Since ListUser.months_in_timespan got changed this assert need repair
-    # assert_equal [3], ListUser.months_in_timespan(firstDay..lastDay).map(&:month)
+    expected = [3]
+    current = months_numbers_in_time_span(first_day, last_day)
+    assert_equal expected, current.flatten.uniq
   end
 
-  test 'getMonthsBetween returns [3, 4, 5] if first day in march and last day in may' do
-    firstDay = Date.new(2012, 3, 31)
-    lastDay = Date.new(2012, 5, 1)
+  test 'months_in_time_span returns [3, 4, 5] if first day in march and last day in may' do
+    first_day = Date.new(2012, 3, 31)
+    last_day = Date.new(2012, 5, 1)
 
-    # TODO: Since ListUser.months_in_timespan got changed this assert need repair
-    # assert_equal [3, 4, 5], ListUser.months_in_timespan(firstDay..lastDay).map(&:month)
+    expected = [3, 4, 5]
+    current = months_numbers_in_time_span(first_day, last_day)
+
+    assert_equal expected, current.flatten.uniq
   end
 
   test 'getMonthsBetween returns correct result timespan overlaps year boundary' do
-    firstDay = Date.new(2011, 3, 3)
-    lastDay = Date.new(2012, 5, 1)
+    first_day = Date.new(2011, 3, 3)
+    last_day = Date.new(2012, 5, 1)
 
-    # TODO: Since ListUser.months_in_timespan got changed this assert need repair
-    # assert_equal (3..12).to_a.concat((1..5).to_a), ListUser.months_in_timespan(firstDay..lastDay).map(&:month)
-  end
-
-  # Set Saturday, Sunday and Wednesday to be a holiday, all others to be a
-  # working day.
-  def defineSaturdaySundayAndWendnesdayAsHoliday
-    Setting['plugin_redmine_workload']['general_workday_monday'] = 'checked'
-    Setting['plugin_redmine_workload']['general_workday_tuesday'] = 'checked'
-    Setting['plugin_redmine_workload']['general_workday_wednesday'] = ''
-    Setting['plugin_redmine_workload']['general_workday_thursday'] = 'checked'
-    Setting['plugin_redmine_workload']['general_workday_friday'] = 'checked'
-    Setting['plugin_redmine_workload']['general_workday_saturday'] = ''
-    Setting['plugin_redmine_workload']['general_workday_sunday'] = ''
-  end
-
-  def assertIssueTimesHashEquals(expected, actual)
-    assert expected.is_a?(Hash), 'Expected is no hash.'
-    assert actual.is_a?(Hash),   'Actual is no hash.'
-
-    assert_equal expected.keys.sort, actual.keys.sort, 'Date keys are not equal'
-
-    expected.keys.sort.each do |day|
-      assert expected[day].is_a?(Hash), "Expected is no hashon day #{day}."
-      assert actual[day].is_a?(Hash),   "Actual is no hash on day #{day}."
-
-      assert expected[day].key?(:hours),      "On day #{day}, expected has no key :hours"
-      assert expected[day].key?(:active),     "On day #{day}, expected has no key :active"
-      assert expected[day].key?(:noEstimate), "On day #{day}, expected has no key :noEstimate"
-      assert expected[day].key?(:holiday),    "On day #{day}, expected has no key :holiday"
-
-      assert actual[day].key?(:hours),        "On day #{day}, actual has no key :hours"
-      assert actual[day].key?(:active),       "On day #{day}, actual has no key :active"
-      assert actual[day].key?(:noEstimate),   "On day #{day}, actual has no key :noEstimate"
-      assert actual[day].key?(:holiday),      "On day #{day}, actual has no key :holiday"
-
-      assert_in_delta expected[day][:hours],   actual[day][:hours], 1e-4, "On day #{day}, hours wrong: #{actual[day][:hours]}"
-      assert_equal expected[day][:active],     actual[day][:active],      "On day #{day}, active wrong"
-      assert_equal expected[day][:noEstimate], actual[day][:noEstimate],  "On day #{day}, noEstimate wrong"
-      assert_equal expected[day][:holiday],    actual[day][:holiday],     "On day #{day}, holiday wrong"
-    end
+    expected = [3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 1, 2]
+    current = months_numbers_in_time_span(first_day, last_day)
+    assert_equal expected, current.flatten.uniq
   end
 
   test 'hours_for_issue_per_day returns {} if time span empty' do
@@ -145,14 +110,17 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 10
     )
 
-    firstDay = Date.new(2013, 5, 31)
-    lastDay = Date.new(2013, 5, 29)
+    first_day = Date.new(2013, 5, 31)
+    last_day = Date.new(2013, 5, 29)
 
-    assertIssueTimesHashEquals({}, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay))
+    assert_issue_times_hash_equals({}, ListUser.send(:hours_for_issue_per_day,
+                                                     issue,
+                                                     first_day..last_day,
+                                                     first_day))
   end
 
   test 'hours_for_issue_per_day works if issue is completely in given time span and nothing done' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     issue = Issue.generate!(
       start_date: Date.new(2013, 5, 31), # A Friday
@@ -161,10 +129,10 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 0
     )
 
-    firstDay = Date.new(2013, 5, 31) # A Friday
-    lastDay = Date.new(2013, 6, 3)   # A Monday
+    first_day = Date.new(2013, 5, 31) # A Friday
+    last_day = Date.new(2013, 6, 3)   # A Monday
 
-    expectedResult = {
+    expected_result = {
       Date.new(2013, 5, 31) => {
         hours: 10.0,
         active: true,
@@ -191,11 +159,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 first_day)
   end
 
   test 'hours_for_issue_per_day works if issue lasts after time span and done_ratio > 0' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     # 30 hours still need to be done, 3 working days until issue is finished.
     issue = Issue.generate!(
@@ -205,10 +177,10 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 25
     )
 
-    firstDay = Date.new(2013, 5, 27) # A Monday, before issue starts
-    lastDay = Date.new(2013, 5, 30) # Thursday, before issue ends
+    first_day = Date.new(2013, 5, 27) # A Monday, before issue starts
+    last_day = Date.new(2013, 5, 30) # Thursday, before issue ends
 
-    expectedResult = {
+    expected_result = {
       # Monday, no holiday, before issue starts.
       Date.new(2013, 5, 27) => {
         hours: 0.0,
@@ -239,11 +211,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 first_day)
   end
 
   test 'hours_for_issue_per_day works if issue starts before time span' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     # 36 hours still need to be done, 2 working days until issue is due.
     # One day has already passed with 10% done.
@@ -254,10 +230,10 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 10
     )
 
-    firstDay = Date.new(2013, 5, 29) # A Wednesday, before issue starts
-    lastDay = Date.new(2013, 6, 1)   # Saturday, before issue ends
+    first_day = Date.new(2013, 5, 29) # A Wednesday, before issue starts
+    last_day = Date.new(2013, 6, 1)   # Saturday, before issue ends
 
-    expectedResult = {
+    expected_result = {
       # Wednesday, holiday, first day of time span.
       Date.new(2013, 5, 29) => {
         hours: 0.0,
@@ -288,11 +264,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 first_day)
   end
 
   test 'hours_for_issue_per_day works if issue completely before time span' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     # 10 hours still need to be done, but issue is overdue. Remaining hours need
     # to be put on first working day of time span.
@@ -303,10 +283,10 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 90
     )
 
-    firstDay = Date.new(2013, 6, 2)  # Sunday, after issue due date
-    lastDay = Date.new(2013, 6, 4)   # Tuesday
+    first_day = Date.new(2013, 6, 2)  # Sunday, after issue due date
+    last_day = Date.new(2013, 6, 4)   # Tuesday
 
-    expectedResult = {
+    expected_result = {
       # Sunday, holiday.
       Date.new(2013, 6, 2) => {
         hours: 0.0,
@@ -330,11 +310,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 first_day)
   end
 
   test 'hours_for_issue_per_day works if issue has no due date' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     # 10 hours still need to be done.
     issue = Issue.generate!(
@@ -344,10 +328,10 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 90
     )
 
-    firstDay = Date.new(2013, 6, 2)  # Sunday
-    lastDay = Date.new(2013, 6, 4)   # Tuesday
+    first_day = Date.new(2013, 6, 2)  # Sunday
+    last_day = Date.new(2013, 6, 4)   # Tuesday
 
-    expectedResult = {
+    expected_result = {
       # Sunday, holiday.
       Date.new(2013, 6, 2) => {
         hours: 0.0,
@@ -371,11 +355,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 first_day)
   end
 
   test 'hours_for_issue_per_day works if issue has no start date' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     # 10 hours still need to be done.
     issue = Issue.generate!(
@@ -385,10 +373,10 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 90
     )
 
-    firstDay = Date.new(2013, 6, 2)  # Sunday
-    lastDay = Date.new(2013, 6, 4)   # Tuesday
+    first_day = Date.new(2013, 6, 2)  # Sunday
+    last_day = Date.new(2013, 6, 4)   # Tuesday
 
-    expectedResult = {
+    expected_result = {
       # Sunday, holiday.
       Date.new(2013, 6, 2) => {
         hours: 0.0,
@@ -412,11 +400,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, firstDay)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 first_day)
   end
 
   test 'hours_for_issue_per_day works if in time span and issue overdue' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     # 10 hours still need to be done, but issue is overdue. Remaining hours need
     # to be put on first working day of time span.
@@ -427,11 +419,11 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 90
     )
 
-    firstDay = Date.new(2013, 5, 30)  # Thursday
-    lastDay = Date.new(2013, 6, 4)    # Tuesday
+    first_day = Date.new(2013, 5, 30)  # Thursday
+    last_day = Date.new(2013, 6, 4)    # Tuesday
     today = Date.new(2013, 6, 2)      # After issue end
 
-    expectedResult = {
+    expected_result = {
       # Thursday, in the past.
       Date.new(2013, 5, 30) => {
         hours: 0.0,
@@ -476,11 +468,15 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, today)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 today)
   end
 
   test 'hours_for_issue_per_day works if issue is completely in given time span, but has started' do
-    defineSaturdaySundayAndWendnesdayAsHoliday
+    define_saturday_sunday_and_wednesday_as_holiday
 
     issue = Issue.generate!(
       start_date: Date.new(2013, 5, 31), # A Friday
@@ -489,11 +485,11 @@ class ListUserTest < ActiveSupport::TestCase
       done_ratio: 0
     )
 
-    firstDay = Date.new(2013, 5, 31) # A Friday
-    lastDay = Date.new(2013, 6, 5)   # A Wednesday
+    first_day = Date.new(2013, 5, 31) # A Friday
+    last_day = Date.new(2013, 6, 5)   # A Wednesday
     today = Date.new(2013, 6, 2)     # A Sunday
 
-    expectedResult = {
+    expected_result = {
       # Friday
       Date.new(2013, 5, 31) => {
         hours: 0.0,
@@ -538,7 +534,11 @@ class ListUserTest < ActiveSupport::TestCase
       }
     }
 
-    assertIssueTimesHashEquals expectedResult, ListUser.send(:hours_for_issue_per_day, issue, firstDay..lastDay, today)
+    assert_issue_times_hash_equals expected_result,
+                                   ListUser.send(:hours_for_issue_per_day,
+                                                 issue,
+                                                 first_day..last_day,
+                                                 today)
   end
 
   test 'hours_per_user_issue_and_day returns correct structure' do
@@ -570,11 +570,13 @@ class ListUserTest < ActiveSupport::TestCase
       project: project2
     )
 
-    firstDay = Date.new(2013, 5, 25)
-    lastDay = Date.new(2013, 6, 4)
+    first_day = Date.new(2013, 5, 25)
+    last_day = Date.new(2013, 6, 4)
     today = Date.new(2013, 5, 31)
 
-    workloadData = ListUser.hours_per_user_issue_and_day(Issue.assigned_to(user).to_a, firstDay..lastDay, today)
+    workloadData = ListUser.hours_per_user_issue_and_day(Issue.assigned_to(user).to_a,
+                                                         first_day..last_day,
+                                                         today)
 
     assert workloadData.key?(user)
 
@@ -679,20 +681,67 @@ class ListUserTest < ActiveSupport::TestCase
     project = Project.generate!
     project.enable_module! :Workload
 
-    projectManagerRole = Role.generate!(name: 'Project manager',
-                                        permissions: [:view_project_workload])
+    project_manager_role = Role.generate!(name: 'Project manager',
+                                          permissions: [:view_project_workload])
 
-    User.add_to_project(user, project, projectManagerRole)
+    User.add_to_project(user, project, project_manager_role)
 
-    projectMember1 = User.generate!
-    User.add_to_project(projectMember1, project)
-    projectMember2 = User.generate!
-    User.add_to_project(projectMember2, project)
+    project_member1 = User.generate!
+    User.add_to_project(project_member1, project)
+    project_member2 = User.generate!
+    User.add_to_project(project_member2, project)
 
     # Create some non-member
     User.generate!
     users = UserSelection.new(user: user)
-    assert_equal [user, projectMember1, projectMember2].map(&:id).sort,
+    assert_equal [user, project_member1, project_member2].map(&:id).sort,
                  users.allowed_to_display.map(&:id).sort
+  end
+
+  private
+
+  def months_numbers_in_time_span(first_day, last_day)
+    DateTools.months_in_time_span(first_day..last_day).map do |span|
+      [span[:first_day].month, span[:last_day].month]
+    end
+  end
+
+  # Set Saturday, Sunday and Wednesday to be a holiday, all others to be a
+  # working day.
+  def define_saturday_sunday_and_wednesday_as_holiday
+    Setting['plugin_redmine_workload']['general_workday_monday'] = 'checked'
+    Setting['plugin_redmine_workload']['general_workday_tuesday'] = 'checked'
+    Setting['plugin_redmine_workload']['general_workday_wednesday'] = ''
+    Setting['plugin_redmine_workload']['general_workday_thursday'] = 'checked'
+    Setting['plugin_redmine_workload']['general_workday_friday'] = 'checked'
+    Setting['plugin_redmine_workload']['general_workday_saturday'] = ''
+    Setting['plugin_redmine_workload']['general_workday_sunday'] = ''
+  end
+
+  def assert_issue_times_hash_equals(expected, actual)
+    assert expected.is_a?(Hash), 'Expected is no hash.'
+    assert actual.is_a?(Hash),   'Actual is no hash.'
+
+    assert_equal expected.keys.sort, actual.keys.sort, 'Date keys are not equal'
+
+    expected.keys.sort.each do |day|
+      assert expected[day].is_a?(Hash), "Expected is no hashon day #{day}."
+      assert actual[day].is_a?(Hash),   "Actual is no hash on day #{day}."
+
+      assert expected[day].key?(:hours),      "On day #{day}, expected has no key :hours"
+      assert expected[day].key?(:active),     "On day #{day}, expected has no key :active"
+      assert expected[day].key?(:noEstimate), "On day #{day}, expected has no key :noEstimate"
+      assert expected[day].key?(:holiday),    "On day #{day}, expected has no key :holiday"
+
+      assert actual[day].key?(:hours),        "On day #{day}, actual has no key :hours"
+      assert actual[day].key?(:active),       "On day #{day}, actual has no key :active"
+      assert actual[day].key?(:noEstimate),   "On day #{day}, actual has no key :noEstimate"
+      assert actual[day].key?(:holiday),      "On day #{day}, actual has no key :holiday"
+
+      assert_in_delta expected[day][:hours],   actual[day][:hours], 1e-4, "On day #{day}, hours wrong: #{actual[day][:hours]}"
+      assert_equal expected[day][:active],     actual[day][:active],      "On day #{day}, active wrong"
+      assert_equal expected[day][:noEstimate], actual[day][:noEstimate],  "On day #{day}, noEstimate wrong"
+      assert_equal expected[day][:holiday],    actual[day][:holiday],     "On day #{day}, holiday wrong"
+    end
   end
 end